[opt] 1.优化代码 2.修改用户列表功能

+++ mkfree-deploy-backend/src/main/java/com/mkfree/deploy/common/PageResult.java
@@ -54,7 +54,7 @@ public class PageResult<T> {
         if (pageSize != null) {
             this.pageSize = pageSize;
         }
-        if (this.pageNo.intValue() > 0) {
+        if (this.pageNo > 0) {
             this.offset = this.pageSize * this.pageNo;
         }
         if (totalCount != null) {
